+++ src/main/java/subway/controller/StationController.java
@@ -4,8 +4,8 @@ import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import subway.domain.Station;
-import subway.request.StationRequest;
-import subway.response.StationResponse;
+import subway.controller.request.StationRequest;
+import subway.controller.response.StationResponse;
 import subway.service.StationService;
 
 import java.net.URI;
