+++ src/main/java/subway/controller/request/SectionRequest.java
@@ -1,4 +1,4 @@
-package subway.request;
+package subway.controller.request;
 
 public class SectionRequest {
     private Long upStationId;
