+++ src/main/java/subway/controller/request/StationRequest.java
@@ -1,4 +1,4 @@
-package subway.request;
+package subway.controller.request;
 
 import subway.domain.Station;
