+++ src/test/java/subway/station/StationAcceptanceTest.java
@@ -8,8 +8,8 @@ import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import subway.AcceptanceTest;
-import subway.request.StationRequest;
-import subway.response.StationResponse;
+import subway.controller.request.StationRequest;
+import subway.controller.response.StationResponse;
 
 import java.util.Arrays;
 import java.util.List;
